+++ login/session.go
@@ -1,4 +1,4 @@
-package usersession
+package login
 
 import (
 	"errors"
